first part of admins calendar

+++ routes/api.php
@@ -84,6 +84,9 @@ Route::group([
     Route::post('role', 'RoleController@update');
     Route::delete('role', 'RoleController@delete');
     
+    Route::get('company/{uid?}/events', 'Api\EventsController@getEvents')
+        ->where('uid', '[a-z0-9]{32}');
+    
 });
 //Route::get('timezones', function(){
 //
